requirements.txt
# mango 🚀 requirements
# Usage: pip install -r requirements.txt

absl-py                 ==2.0.0
cachetools              ==5.3.2
certifi                 ==2023.11.17
charset-normalizer      ==3.3.2
colorama                ==0.4.6
contourpy               ==1.2.0
cycler                  ==0.12.1
et-xmlfile              ==1.1.0
filelock                ==3.13.1
fonttools               ==4.45.1
google-auth             ==2.23.4
google-auth-oauthlib    ==1.1.0
grpcio                  ==1.59.3
h5py                    ==3.10.0
idna                    ==3.6
imageio                 ==2.34.0
importlib-metadata      ==6.8.0
importlib-resources     ==6.1.1
Jinja2                  ==3.1.2
kiwisolver              ==1.4.5
lazy_loader             ==0.3
Markdown                ==3.5.1
MarkupSafe              ==2.1.3
matplotlib              ==3.8.2
mpmath                  ==1.3.0
networkx                ==3.2.1
numpy                   ==1.26.4
oauthlib                ==3.2.2
opencv-python           ==4.8.1.78
openpyxl                ==3.1.2
packaging               ==23.2
pandas                  ==2.2.1
Pillow                  ==10.1.0
pip                     ==23.3.1
protobuf                ==4.23.4
psutil                  ==5.9.8
py-cpuinfo              ==9.0.0
pyasn1                  ==0.5.1
pyasn1-modules          ==0.3.0
pyparsing               ==3.1.1
PySide6                 ==6.6.3
PySide6_Addons          ==6.6.3
PySide6_Essentials      ==6.6.3
python-dateutil         ==2.8.2
pytz                    ==2024.1
PyYAML                  ==6.0.1
requests                ==2.31.0
requests-oauthlib       ==1.3.1
rsa                     ==4.9
scikit-image            ==0.22.0
scipy                   ==1.11.4
seaborn                 ==0.13.2
setuptools              ==68.0.0
shiboken6               ==6.6.3
six                     ==1.16.0
sympy                   ==1.12
tensorboard             ==2.15.1
tensorboard-data-server ==0.7.2
thop                    ==0.1.1.post2209072238
tifffile                ==2024.2.12
torch                   ==2.0.0+cu118
torchaudio              ==2.0.1+cu118
torchvision             ==0.15.1+cu118
tqdm                    ==4.66.1
typing_extensions       ==4.8.0
tzdata                  ==2024.1
ultralytics             ==8.1.33
urllib3                 ==2.1.0
Werkzeug                ==3.0.1
wheel                   ==0.41.2
zipp                    ==3.17.0
